♻️ is_philo_dieを、can_stop_threadに変更
