+++ philo/src/init_die_judge.c
@@ -6,7 +6,7 @@
 /*   By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/19 19:35:53 by miyuu             #+#    #+#             */
-/*   Updated: 2025/04/20 18:55:58 by miyuu            ###   ########.fr       */
+/*   Updated: 2025/04/20 19:39:32 by miyuu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,9 +21,9 @@ void	print_one_die_judge(t_die_judge *judge)
 	{
 		printf("  [%d] %ld (addr: %p)\n", i, judge->last_eat_time[i], (void *)&judge->last_eat_time[i]);
 	}
-	printf("is_philo_die    : %s (addr: %p)\n",
-		*(judge->is_philo_die) ? "true" : "false",
-		(void *)judge->is_philo_die);
+	printf("can_stop_thread    : %s (addr: %p)\n",
+		*(judge->can_stop_thread) ? "true" : "false",
+		(void *)judge->can_stop_thread);
 	printf("is_eat_full   :\n");
 	for (int i = 0; i < judge->u_rules.total_philo; i++)
 	{
@@ -38,7 +38,7 @@ void	init_die_judge(t_die_judge	*die_judge, t_univ_rules rules, \
 	die_judge->start_tv_ms = start_tv_ms;
 	die_judge->u_rules = rules;
 	die_judge->last_eat_time = s_data->last_eat_time;
-	die_judge->is_philo_die = s_data->is_philo_die;
+	die_judge->can_stop_thread = s_data->can_stop_thread;
 	die_judge->is_eat_full = s_data->is_eat_full;
 	print_one_die_judge(die_judge);
 }
