+++ philo/src/init_thread_arg.c
@@ -6,7 +6,7 @@
 /*   By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/15 20:32:11 by miyuu             #+#    #+#             */
-/*   Updated: 2025/04/20 18:19:51 by miyuu            ###   ########.fr       */
+/*   Updated: 2025/04/20 19:39:32 by miyuu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ void	print_one_thread_arg(t_thread_arg *arg)
 	printf("second_fork_n  : %d\n", arg->second_fork_n);
 	printf("start_tv_ms   : %ld\n", arg->start_tv_ms);
 	printf("last_eat_time : %ld (%p)\n", *(arg->last_eat_time), (void *)arg->last_eat_time);
-	printf("is_philo_die  : %s (%p)\n", *(arg->is_philo_die) ? "true" : "false", (void *)arg->is_philo_die);
+	printf("can_stop_thread  : %s (%p)\n", *(arg->can_stop_thread) ? "true" : "false", (void *)arg->can_stop_thread);
 	printf("is_eat_full : %s (%p)\n", *(arg->is_eat_full) ? "true" : "false", (void *)arg->is_eat_full);
 	printf("----------------------\n");
 }
@@ -61,7 +61,7 @@ void	init_thread_arg(t_univ_rules rules, t_share_data *s_data, \
 		arg[i].u_rules = rules;
 		arg[i].start_tv_ms = start_tv_ms;
 		arg[i].last_eat_time = &s_data->last_eat_time[i];
-		arg[i].is_philo_die = s_data->is_philo_die;
+		arg[i].can_stop_thread = s_data->can_stop_thread;
 		arg[i].is_eat_full = &s_data->is_eat_full[i];
 		assign_forks(&arg[i], s_data->forks, rules.total_philo, i);
 		print_one_thread_arg(&arg[i]);
